requirements.txt
alembic==0.9.9
Babel==2.5.3
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
dominate==2.3.1
elasticsearch==6.2.0
Flask==1.0
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
PyJWT==1.6.1
PyMySQL==0.8.0
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
